docs: document docker_wait_for_services_timeout
Document `docker_wait_for_services_timeout` variable.

I was setting this option using `extra_configs` because I couldn't find a dedicated variable to set. However, doing so will actually cause a change on every run because the default will be re-written just for the extra config to correct it again.